# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Navarchos PdM API",
    "version": "1.0.0"
  },
  "paths": {
    "/data": {
      "post": {
        "description": "Returns the results of the PdM algorithm",
        "consumes": [
            "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Input",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "timestamp": {
                  "type": "string",
                  "description": "Timestamp of the data",
                  "example": "2024-02-22 01:14:34"
                },
                "features": {
                  "type": "array",
                  "description": "A json array of floats for the features",
                  "example": "[0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0, 1.1, 1.2, 1.3, 1.4, 1.5]"
                },
                "source": {
                  "type": "string",
                  "description": "Source",
                  "example": "4"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "source": {
                  "type": "string",
                  "description": "Source",
                  "example": "4"
                },
                "timestamp": {
                  "type": "string",
                  "description": "Timestamp of the data",
                  "example": "2024-02-22 01:14:34"
                },
                "alarm": {
                  "type": "boolean",
                  "description": "Whether an alarm is raised",
                  "example": "2024-02-22 01:14:34True"
                },
                "scores": {
                  "type": "array",
                  "description": "A json array of floats for the scores of each feature",
                  "example": "[0,0,0,0,0,0,0,0,0,0,0,0,0,2,0]"
                },
                "thresholds": {
                  "type": "array",
                  "description": "A json array of floats for the thresholds of each features",
                  "example": "[1,1,1,1,1,1,1,1,1,1,1,1,1,1,1]"
                },
                "description": {
                  "type": "string",
                  "description": "Description",
                  "example": "<12>"
                }
              }
            }
          },
          "406": {
            "description": "Missing parameters on input"
          },
          "500": {
            "description": "Error during the PdM process"
          }
        }
      }
    },
    "/event": {
        "post": {
          "description": "Insert an event to the PdM algorithm",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Input",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "timestamp": {
                    "type": "string",
                    "description": "Timestamp of the data",
                    "example": "2024-02-22 01:14:34"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the event",
                    "example": "mpla mpla"
                  },
                  "source": {
                    "type": "string",
                    "description": "Source",
                    "example": "4"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Event ingested"
            },
            "406": {
              "description": "Missing parameters on input"
            },
            "500": {
              "description": "Error during the ingestion"
            }
          }
        }
      }
  }
}